Fix authentication logic.
My previous commit had the bug that "$authenticated" was a mere counter,
increased each time a valid auth directive was found. This was wrong.
The current commit maintains two flags; a configuration file is only
taken in to account if authentication is enabled and a valid user is
found, where the last found auth directive in the AWStats config file
overrides all previous ones (as should be).